% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/groupGeneType.R
\docType{data}
\name{ncRNA}
\alias{ncRNA}
\title{groupGeneType}
\format{\preformatted{ chr [1:5] "sense_intronic" "3prime_overlapping_ncrna" "processed_transcript" "sense_overlapping" "Other_lncRNA"
}}
\usage{
ncRNA
}
\arguments{
\item{x1:}{\code{geneType}}
}
\value{
y1: new group name
}
\description{
This function grouped ensembl gene types to bigger gene group type
}
\keyword{keywords}

